+++ src/drto/__init__.py
@@ -32,6 +32,7 @@ from drto.info import Info, info
 from drto.initialize_steady_state import SteadyStateInitReport, initialize_steady_state
 from drto.advanced_step import advanced_step_controller
 from drto.cold_start import ColdStartReport, cold_start_dynamic
+from drto.warm_start import WarmStartReport, warm_start_dynamic
 from drto.objective import build_objective
 from drto.plotting import plot_controls, plot_stage_cost, plot_states
 
@@ -55,6 +56,8 @@ __all__ = [
     "build_objective",
     "advanced_step_controller",
     "cold_start_dynamic",
+    "warm_start_dynamic",
+    "WarmStartReport",
     "ColdStartReport",
     "plot_states",
     "plot_controls",